requirements.txt
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.2
asgiref==3.7.2
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
autoflake==1.4
Babel==2.9.1
backcall==0.2.0
bandit==1.7.4
beautifulsoup4==4.10.0
black==20.8b1
blinker==1.7.0
cachetools==5.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
cloudpickle==3.0.0
coverage==5.5
darglint==1.8.1
dask==2024.2.0
dask-glm==0.3.2
dask-ml==2023.3.24
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
distributed==2024.2.0
Django==5.0.2
djangorestframework==3.14.0
doc8==0.8.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.4
eradicate==1.0
executing==0.8.3
filelock==3.6.0
flake8==3.9.2
flake8-2020==1.6.1
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
flask==3.0.0
fsspec==2024.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-core==2.3.1
google-cloud-datastore==2.7.2
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
huggingface-hub==0.20.3
idna==3.3
imagesize==1.3.0
importlib-metadata==7.0.1
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.1.1
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.71
jupyter-client==7.1.2
jupyter-core==4.9.2
llvmlite==0.42.0
locket==1.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
msgpack==1.0.7
multipledispatch==1.0.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numba==0.59.0
numpy==1.26.4
packaging==21.3
pandas==2.2.0
parso==0.8.3
partd==1.4.1
pathspec==0.9.0
pbr==5.8.1
pep8-naming==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
pipwin==0.5.2
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.1
Pygments==2.11.2
pyjsparser==2.7.1
pyparsing==3.0.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dev-tools==2020.9.10
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyupgrade==2.31.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
restructuredtext-lint==1.4.0
rsa==4.8
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sparse==0.15.1
Sphinx==2.4.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.2.0
stevedore==3.5.0
tblib==3.0.0
testfixtures==6.18.5
threadpoolctl==3.3.0
tokenize-rt==4.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.1
tox==3.24.5
tox-travis==0.12
tqdm==4.66.2
traitlets==5.1.1
typed-ast==1.5.2
typing-extensions==4.9.0
tzdata==2024.1
tzlocal==4.1
urllib3==1.26.9
virtualenv==20.14.0
wcwidth==0.2.5
wemake-python-styleguide==0.14.1
werkzeug==3.0.1
zict==3.0.0
zipp==3.17.0
